Module Drom_lib__Configuration_Parser

Parsing configuration files.

Configuration files in the form of INI files that were common under Microsoft systems have grown popular in the open source world, and are used my some major pieces of software, like Subversion or GIT.

We provide a parser to analyse these files. This parser is a functor parameterised by handlers for the events found in a configuration file. These events are of the following kinds:

comments;

sections;

bindings;

parse errors.

type pos = Stdlib.Lexing.position

Position in an input stream.

type excerpt

The type of excerpts.

val startpos : excerpt -> pos

The start position of an excerpt.

val endpos : excerpt -> pos

The end position of an excerpt.

val text : excerpt -> string

The text of an excerpt.

type error =

| Illegal_character

| Illegal_escape_sequence

| Unterminated_string

| Expecting_binding

| Expecting_value

| Expecting_path_elt

| Expecting_path_sep_or_term

Parsing errors.

val error_to_string : error -> string

Textual description of errors.

module type Definition = sig ... end

The input signature of the functor Configuration_Parser.Make.

module type S = sig ... end

The output signature of the functor Configuration_Parser.Make.

module Make : functor (D : Definition) -> S with type t = D.t

Functor building an implementation of the configuration file parser given a parser definition.