Module Drom_lib__VersionCompare

Version comparison function used throughout. From the Dose suite.

Comparing debian version strings

val equal : string -> string -> bool

returns

true iff the two strings define the same version. Hence, the result may be true even when the two string differ syntactically.

val compare : string -> string -> int

compare x y returns 0 if x is eqivalent to y, -1 if x is smaller than y, and 1 if x is greater than y. This is consistent with Stdlib.compare.